+++ pkg/hedgedmetrics/metrics_test.go
@@ -1,67 +1,46 @@
 package hedgedmetrics
 
 import (
+	"math"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/cristalhq/hedgedhttp"
+	"github.com/grafana/tempo/pkg/util/test"
 	"github.com/prometheus/client_golang/prometheus"
-	dto "github.com/prometheus/client_model/go"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
 )
 
-// MockCounter satisfies prometheus.Counter
-type MockCounter struct {
-	mock.Mock
-	val float64
+func TestDiffCounter(t *testing.T) {
+	ctr := prometheus.NewCounter(prometheus.CounterOpts{Name: test.RandomString()})
+	dc := &diffCounter{previous: 0, counter: ctr}
+
+	dc.addAbsoluteToCounter(5)
+	require.Equal(t, 5.0, ctrVal(t, ctr))
+
+	dc.addAbsoluteToCounter(7)
+	require.Equal(t, 7.0, ctrVal(t, ctr))
+
+	dc.addAbsoluteToCounter(57)
+	require.Equal(t, 57.0, ctrVal(t, ctr))
 }
 
-func (m *MockCounter) Desc() *prometheus.Desc {
-	return nil
-}
+func TestDiffCounterOverflow(t *testing.T) {
+	ctr := prometheus.NewCounter(prometheus.CounterOpts{Name: test.RandomString()})
+	dc := &diffCounter{previous: 0, counter: ctr}
 
-func (m *MockCounter) Write(_ *dto.Metric) error {
-	return nil
-}
+	// start with something large
+	bigNum := math.MaxUint64 - uint64(1)
+	dc.addAbsoluteToCounter(bigNum)
+	require.Equal(t, float64(bigNum), ctrVal(t, ctr))
 
-func (m *MockCounter) Describe(_ chan<- *prometheus.Desc) {
-	// no-op
-}
+	// then wrap the uint64s
+	dc.addAbsoluteToCounter(2)
+	require.Equal(t, 1.0, ctrVal(t, ctr)) // this is one off b/c of the internal counter wrap in prom. this could be corrected for but doesn't matter
 
-func (m *MockCounter) Collect(_ chan<- prometheus.Metric) {
-	// no-op
-}
-
-func (m *MockCounter) Add(v float64) {
-	m.val += v
-	m.Called(v)
-}
-
-func (m *MockCounter) Inc() {
-	m.Add(1.0)
-}
-
-func TestCounterWithValue(t *testing.T) {
-	mockCounter := new(MockCounter)
-	mockCounter.On("Add", mock.AnythingOfType("float64")).Return()
-
-	counter := NewCounterWithValue(mockCounter)
-
-	assert.Equal(t, int64(0), counter.Value())
-
-	counter.Add(5)
-	assert.Equal(t, int64(5), counter.Value())
-	mockCounter.AssertCalled(t, "Add", 5.0)
-
-	counter.Add(3)
-	assert.Equal(t, int64(8), counter.Value())
-	mockCounter.AssertCalled(t, "Add", 3.0)
-
-	counter.Add(0)
-	assert.Equal(t, int64(8), counter.Value())
-	mockCounter.AssertCalled(t, "Add", 0.0)
+	dc.addAbsoluteToCounter(7)
+	require.Equal(t, 6.0, ctrVal(t, ctr))
 }
 
 // MockStatsProvider is StatsProvider for testing
@@ -88,39 +67,42 @@ func (m *MockStatsProvider) SetStats(actual, requested uint64) {
 }
 
 func TestPublish(t *testing.T) {
-	promCounter := prometheus.NewCounter(prometheus.CounterOpts{
-		Namespace: "test",
-		Name:      "counter",
-		Help:      "test counter",
-	})
-	counter := NewCounterWithValue(promCounter)
+	ctr := prometheus.NewCounter(prometheus.CounterOpts{Name: test.RandomString()})
 	stats := &MockStatsProvider{}
 
-	// start the Publishing every 10ms
-	Publish(stats, counter, 10*time.Millisecond)
+	publishWithDuration(stats, ctr, 10*time.Millisecond)
 
-	assert.Equal(t, int64(0), counter.Value())
+	require.Equal(t, 0.0, ctrVal(t, ctr))
 
 	// Set initial stats values
 	stats.SetStats(5, 5)
 	time.Sleep(30 * time.Millisecond)
-	assert.Equal(t, int64(0), counter.Value())
+	require.Equal(t, 0.0, ctrVal(t, ctr))
 
 	stats.SetStats(15, 10)
 	time.Sleep(30 * time.Millisecond)
-	assert.Equal(t, int64(5), counter.Value())
+	require.Equal(t, 5.0, ctrVal(t, ctr))
 
 	stats.SetStats(28, 20)
 	time.Sleep(30 * time.Millisecond)
-	assert.Equal(t, int64(8), counter.Value())
+	require.Equal(t, 8.0, ctrVal(t, ctr))
 
 	stats.SetStats(38, 25)
 	time.Sleep(30 * time.Millisecond)
-	assert.Equal(t, int64(13), counter.Value())
+	require.Equal(t, 13.0, ctrVal(t, ctr))
 
 	time.Sleep(30 * time.Millisecond)
 
 	// counter doesn't increase if stats stay same
 	time.Sleep(30 * time.Millisecond)
-	assert.Equal(t, int64(13), counter.Value())
+	require.Equal(t, 13.0, ctrVal(t, ctr))
+}
+
+func ctrVal(t *testing.T, ctr prometheus.Counter) float64 {
+	t.Helper()
+
+	val, err := test.GetCounterValue(ctr)
+	require.NoError(t, err)
+
+	return val
 }