+++ tempodb/backend/instrumentation/hedged_requests.go
@@ -7,18 +7,15 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promauto"
 )
 
-var (
-	hedgedRequestsMetrics = promauto.NewCounter(
-		prometheus.CounterOpts{
-			Namespace: "tempodb",
-			Name:      "backend_hedged_roundtrips_total",
-			Help:      "Total number of hedged backend requests.",
-		},
-	)
-	hedgedRequestsMetricsWithValue = hedgedmetrics.NewCounterWithValue(hedgedRequestsMetrics)
+var hedgedRequestsMetrics = promauto.NewCounter(
+	prometheus.CounterOpts{
+		Namespace: "tempodb",
+		Name:      "backend_hedged_roundtrips_total",
+		Help:      "Total number of hedged backend requests.",
+	},
 )
 
 // PublishHedgedMetrics flushes metrics from hedged requests every 10 seconds
 func PublishHedgedMetrics(s *hedgedhttp.Stats) {
-	hedgedmetrics.Publish(s, hedgedRequestsMetricsWithValue, hedgedmetrics.PublishDuration)
+	hedgedmetrics.Publish(s, hedgedRequestsMetrics)
 }